refactor: Overhaul map view into MaxExplorer
Add feature for user to drop markers with option to drag, remove, and
save dropped markers.

Add circle marker for user's current position.

Add nicer style for search menu that looks embedded within MapPane.

Remove redundant components
