--- client/src/components/Explore/style.css
@@ -1,5 +1,0 @@
-.fullscreen {
-	height: inherit;
-	width: inherit;
-	display: block;
-}
