+++ client/src/components/LeafletOsmMap/index.js
@@ -1,17 +1,21 @@
-import { h, Component } from 'preact';
+import { h, Component } from "preact";
+import style from "./style";
+import MapPane from './MapPane';
+
+/**
+ * Leaflet related imports: leaflet, pouchdb module, and routing machine module
+ */
 import '../../../node_modules/leaflet/dist/leaflet.css';
+import '../../../node_modules/leaflet-routing-machine/dist/leaflet-routing-machine.css';
 import L from '../../js/leaflet-tileLayer-pouchdb-cached';
-
-
-// // TODO - remove this module from dependencies
-// import Map, { GoogleApiWrapper, Marker, InfoWindow } from 'google-maps-react';
+import Routing from '../../../node_modules/leaflet-routing-machine/src/index.js';
 
 /**
  * TIle layer configuration and attribution constants
  */
 const OSM_URL = 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png';
 const OSM_ATTRIB = '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
-const TILE_LAYER = new L.TileLayer(OSM_URL, {
+const OSM_TILE_LAYER = new L.TileLayer(OSM_URL, {
     attribution: OSM_ATTRIB,
     useCache: true,
     crossOrigin: true,
@@ -20,67 +24,126 @@ const TILE_LAYER = new L.TileLayer(OSM_URL, {
 // redirect marker icon path to assets directory
 L.Icon.Default.imagePath = '../../assets/icons/leaflet/';
 
-export default class MapContainer extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            map: null,
-            lat: null,
-            lng: null,
-            watchID: null,
-            screenHeight: screen.height * 0.95,
-        }
-        this.initMap = this.initMap.bind(this);
-      }
-
-      componentDidMount() {
-        // initialize map container if null
-        if (!this.state.map) {
-            this.setState({
-                map: new L.Map('map'),
-            });
-        }
-
-        // continuously track users position
-        const watchID = navigator.geolocation.watchPosition(this.initMap);
-        this.setState({
-            watchID: watchID,
-        });
-      }
-    
-      initMap(position) {
-        // don't initialize map if it already has OSM tile layer
-        if (this.state.map.hasLayer(TILE_LAYER)) {
-            return;
-        }
-
-        this.setState({
-            lat: position.coords.latitude || 51.3,
-            lng: position.coords.longitude || 0.7,
-        });
-    
-        this.state.map.setView(L.latLng(this.state.lat, this.state.lng), 15);
-        this.state.map.addLayer(TILE_LAYER);
-    
-        L.marker([this.state.lat, this.state.lng])
-          .addTo(this.state.map).bindPopup('Current Location').openPopup();
-      }
-
-    render() {
-        const styles = {
-            height: this.state.screenHeight,  // HAVE TO SET HEIGHT TO RENDER MAP
-        }
-        return (
-            <div id="map" style={styles}/>
-        )
-    };
-
-    componentWillUnmount() {
-        console.log('begin - componentWillUnmount() ', this.state.map);
-        // stop watching user position when map is unmounted
-        navigator.geolocation.clearWatch(this.state.watchID);
-        this.state.map.remove();
-        console.log('begin - componentWillUnmount() ', this.state.map);
+export default class LeafletOSMMap extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      map: null,
+      mapCenter: null,
+      userMarker: null,
     }
+    this.onLocationFound = this.onLocationFound.bind(this);
+    this.onLocationError = this.onLocationError.bind(this);
+    this.onMapClick = this.onMapClick.bind(this);
+  }
+
+  componentDidMount() {
+    // initialize map container if null
+    if (!this.state.map) {
+      this.setState({
+          map: L.map('map', {
+            zoomControl: false,
+            zoom: 16,
+          }),
+      });
+    }
+    // add zoom to bottom left
+    const zoomControl = L.control.zoom().setPosition('bottomleft').addTo(this.state.map);
+
+    this.state.map.addLayer(OSM_TILE_LAYER);
+
+    // attempt to get user's current location via device
+    this.state.map.locate({
+      setView: true,
+      enableHighAccuracy: true,
+      maxZoom: 16,
+    });
+
+    // configure map events
+    this.state.map.on('locationfound', this.onLocationFound);
+    this.state.map.on('locationerror', this.onLocationError);
+    this.state.map.on('click', this.onMapClick);
+  }
+
+  /**
+   * Handle leaflet map get device location event
+   * @param {*} event 
+   */
+  onLocationFound(event) {
+    const userMarker = L.circleMarker(event.latlng, {
+      radius: 8,
+      weight: 3,
+      fillColor: 'red',
+    }).addTo(this.state.map)
+      .bindPopup('You Are Here');
+
+    this.setState({
+      mapCenter: event.latlng,
+      userMarker: userMarker,
+    });
+    console.log('onLocationFound() - check state.map: ', this.state.map);
+  }
+
+  /**
+   * Handle leaflet map get device location failure
+   * @param {*} event 
+   */
+  onLocationError(event) {
+    // TODO - dummy message for now if don't have permission to get user
+    // user location. Next step is to notify and request permission again.
+    alert(event.message);
+  }
+
+  /**
+   * On map click event, add a marker to the map at the clicked location.
+   *
+   * @param {*} event 
+   */
+  onMapClick(event) {
+    const droppedPin = L.marker(event.latlng, {
+      draggable: true,
+      autoPan: true,
+    }).addTo(this.state.map);
+
+    const container = L.DomUtil.create('div');
+    const saveBtn = createButton('Save', container);
+    const deleteBtn = createButton('Remove', container);
+  
+    L.DomEvent.on(saveBtn, 'click', function() {
+      // TODO - save marker location to db if possible or queue to
+      // save when backend server is available
+      alert(`Save marker at: ${event.latlng}!`);
+    });
+  
+    L.DomEvent.on(deleteBtn, 'click', function() {
+      droppedPin.remove();
+    });
+
+    droppedPin.bindPopup(container);
+    console.log('onLocationFound() - check state.map: ', this.state.map);
+  }
+
+  render() {
+    return (
+      <div class={style.fullscreen}>
+        <MapPane height={screen.height * 0.95}/>
+      </div>
+    );
+  }
+
+  componentWillUnmount() {
+    console.log('begin - componentWillUnmount() ', this.state.map);
+    // stop watching user position when map is unmounted
+    // map.stopWatch();
+    this.state.map.remove();
+    this.setState({map: null});
+    console.log('end - componentWillUnmount() ', this.state.map);
+  }
 }
 
+function createButton(label, container) {
+  var btn = L.DomUtil.create('button', '', container);
+  btn.setAttribute('type', 'button');
+  btn.innerHTML = label;
+  return btn;
+}
