--- client/src/components/search/style.css
@@ -1,10 +1,0 @@
-.hide {
-  display: none;
-}
-
-.search {
-  width: 100%;
-  height: 36px;
-  text-align: center;
-  background-color: aliceblue;
-}
