+++ client/src/routes/maps/index.js
@@ -1,18 +1,17 @@
 import { h, Component } from "preact";
 import style from "./style";
-// import MapContainer from '../../components/LeafletOsmMap';
-import Search from '../../components/AppSearch'
+import MapContainer from '../../components/LeafletOsmMap';
+import Search from '../../components/Search'
 import SearchResults from '../../components/SearchResults';
-import Explore from '../../components/Explore';
 
-export default class Maps extends Component {
+export default class MapExplorer extends Component {
   render() {
     return (
       <div class={style.maps}>
         <Search>
           <SearchResults />
         </Search>
-        <Explore />
+        <MapContainer />
       </div>
     );
   }
